Separate CSS-in-JS from the component files
