+++ src/components/Checkbox/Checkbox.theme.tsx
@@ -1,0 +1,10 @@
+import { styled } from "@mui/system";
+
+export const Label = styled("label")({
+  display: "block",
+  paddingTop: "5px",
+  paddingBottom: "5px",
+  whiteSpace: "nowrap",
+  overflow: "hidden",
+  textOverflow: "ellipsis",
+});
