+++ src/components/CheckboxList/CheckboxList.theme.tsx
@@ -1,0 +1,11 @@
+import { Button } from "@mui/material";
+import { styled } from "@mui/system";
+
+export const SelectAllButton = styled(Button)({
+  marginTop: "15px",
+});
+
+export const InputContainer = styled("div")({
+  paddingTop: "10px",
+  paddingBottom: "10px",
+});
