+++ src/components/FetchAlert/FetchAlert.theme.tsx
@@ -1,0 +1,9 @@
+import { styled } from "@mui/system";
+
+export const Container = styled("div")({
+  textAlign: "center",
+  position: "absolute",
+  left: "50%",
+  top: "45%",
+  transform: "translate(-50%, -50%)",
+});
