+++ src/components/Filters/Filters.theme.tsx
@@ -1,0 +1,10 @@
+import { styled } from "@mui/system";
+import { Drawer } from "@mui/material";
+
+export const StyledDrawer = styled(Drawer)({
+  "& .MuiDrawer-paper": {
+    paddingLeft: "40px",
+    paddingRight: "40px",
+    width: "340px",
+  },
+});
