+++ src/components/Header/Header.theme.tsx
@@ -1,0 +1,10 @@
+import { styled } from "@mui/system";
+import { Icon, Toolbar } from "@mui/material";
+
+export const StyledToolbar = styled(Toolbar)({
+  justifyContent: "space-between",
+});
+
+export const FilterIcon = styled(Icon)({
+  cursor: "pointer",
+});
