+++ src/components/Paging/Paging.theme.tsx
@@ -1,0 +1,7 @@
+import { styled } from "@mui/system";
+import { Pagination } from "@mui/material";
+
+export const StyledPagination = styled(Pagination)({
+  display: "flex",
+  justifyContent: "center",
+});
