+++ src/components/PlanetList/PlanetList.theme.tsx
@@ -1,0 +1,14 @@
+import { styled } from "@mui/system";
+
+export const Container = styled("main")({
+  paddingTop: "100px",
+});
+
+export const List = styled("div")({
+  display: "flex",
+  flexDirection: "row",
+  justifyContent: "center",
+  flexWrap: "wrap",
+  paddingTop: "20px",
+  paddingBottom: "20px",
+});
